lin_res_rangga/requirements.txt
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
cycler==0.11.0
Flask==1.1.2
fonttools==4.29.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.1.0
jwcrypto==0.8
kiwisolver==1.3.2
MarkupSafe==1.1.1
matplotlib==3.5.1
mysql-connector-python==8.0.22
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.13.0
pycparser==2.20
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==1.0.1
